Feature: Added lerp method to vector class, updated tests to break up AlgebraSpec, and updated Vector docs page
